Extrayendo componente del screen de Character para mayor limpieza
